pyproject.toml
[project]
name = "sample-florence2-object-detection"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "einops>=0.8.1",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "requests>=2.32.4",
    "timm>=1.0.19",
    "torch>=2.8.0",
    "transformers==4.49.0",
]
